test/inspector/debugger/wasm-gc-breakpoints-expected.txt

Tests GC object inspection.

Running test: test
Instantiating.
Waiting for wasm script (ignoring first non-wasm script).
Setting breakpoint at offset 100 on script wasm://wasm/00670fd6
Calling main()
Paused:
Script wasm://wasm/00670fd6 byte offset 100: Wasm opcode 0x21 (kExprLocalSet)
Scope:
at $main (0:100):
 - scope (wasm-expression-stack):
   0: Array ((ref $ArrC))
     object details:
   0: Struct ((ref null $StrA))
   length: 1 (number)
 - scope (local):
   $varA: Struct ((ref null $StrA))
   $varB: null ((ref null $ArrC))
     object details:
   $byte: 127 (i8)
   $word: 32767 (i16)
   $pointer: Struct ((ref $veryLongNameWithMoreThanOneHundredAndTwentyEightCharactersToTestThatWeAreHandlingStringBufferOverflowWithoutCrashing_ThisWontGetTruncated))
 - scope (module):
   instance: exports: "main" (Function)
   module: Module
   functions: "$main": (Function)
   globals: "$global0": function $main() { [native code] } ((ref null $type3))
at (anonymous) (0:17):
   -- skipped
exports.main returned!
